+++ src/WebJobs.Extensions.CosmosDB/release_notes.md
@@ -1,3 +1,5 @@
-## Microsoft.Azure.WebJobs.Extensions.CosmosDB 4.12.0
+## Microsoft.Azure.WebJobs.Extensions.CosmosDB 4.13.0
 
-- Update Microsoft.Azure.Cosmos to 3.54.0 (#969)
+- Update Microsoft.Azure.Cosmos to 3.55.0 (#976)
+
+**NOTE**: The stable version of this package does not include CosmosDB preview features. Use the `-preview` versions of this package for CosmosDB preview service features.